requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
lxml==4.9.2
markdownify==0.11.6
mysqlclient==2.1.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
